Crate polkadot_node_core_pvf_common

Contains functionality related to PVFs that is shared by the PVF host and the PVF workers.

Re-exports

pub use sp_tracing;

Modules

error

execute

executor_intf

Interface to the Substrate Executor

prepare

pvf

tests

worker

Functionality common to both prepare and execute workers.

Macros

decl_worker_main

Use this macro to declare a fn main() {} that will create an executable that can be used for spawning the desired worker.

Structs

ProcessTime

CPU Time Used by The Whole Process

Functions

framed_recv

Read some data prefixed by its length from r.

framed_send

Write some data prefixed by its length into w.