polkadot_node_core_pvf_common

Module execute

Structs

ExecuteRequest

A request to execute a PVF

Handshake

The payload of the one-time handshake that is done when a worker process is created. Carries data from the host to the worker.

WorkerResponse

The response from the execution worker.

Enums

JobError

An unexpected error occurred in the execution job process. Because this comes from the job, which executes untrusted code, this error must likewise be treated as untrusted. That is, we cannot raise an internal error based on this.

JobResponse

The successful response from a job on the execution worker.

WorkerError

An error occurred in the worker process.

Type Aliases

JobResult

The result of a job on the execution worker.