Module polkadot_node_core_pvf_common::executor_interface

Interface to the Substrate Executor

Constants

DEFAULT_CONFIG

Functions

create_runtime_from_artifact_bytes⚠

Constructs the runtime for the given PVF, given the artifact bytes.

execute_artifact⚠

Executes the given PVF in the form of a compiled artifact and returns the result of execution upon success.

params_to_wasmtime_semantics

prepare

Runs preparation on the given runtime blob. If successful, it returns a serialized compiled artifact which can then be used to pass into Executor::execute after writing it to the disk.

prevalidate

Runs the prevalidation on the given code. Returns a RuntimeBlob if it succeeds.