Module polkadot_node_core_pvf_common::worker

Functionality common to both prepare and execute workers.

Modules

security

Functionality for securing workers.

thread

Functionality related to threads spawned by the workers.

Structs

PipeFd

A wrapper around a file descriptor used to encapsulate and restrict functionality for pipe operations.

WorkerInfo

Enums

WorkerKind

Constants

JOB_TIMEOUT_OVERHEAD

Some allowed overhead that we account for in the “CPU time monitor” thread’s sleeps, on the child process.

Functions

cpu_time_monitor_loop

Loop that runs in the CPU time monitor thread on prepare and execute jobs. Continuously wakes up and then either blocks for the remaining CPU time, or returns if we exceed the CPU timeout.

pipe2_cloexec

run_worker

Initializes the worker process, then runs the given event loop, which spawns a new job process to securely handle each incoming request.

stringify_panic_payload

Attempt to convert an opaque panic payload to a string.