Module polkadot_node_core_pvf_common::worker::security::seccomp

Functionality for sandboxing workers by restricting their capabilities by blocking certain syscalls with seccomp.

For security we block the following:

creation of new sockets - these are unneeded in PVF jobs, and we can safely block them without affecting consensus.

io_uring - allows for networking and needs to be blocked. See below for a discussion on the safety of doing this.

Safety of blocking io_uring

io_uring is just a way of issuing system calls in an async manner, and there is nothing stopping wasmtime from legitimately using it. Fortunately, at the moment it does not. Generally, not many applications use io_uring in production yet, because of the numerous kernel CVEs discovered. It’s still under a lot of development. Android outright banned io_uring for these reasons.

Considering io_uring’s status discussed above, and that it very likely would get detected either by our static analysis or by testing, we think it is safe to block it.

Consensus analysis

If execution hits an edge case code path unique to a given machine, it’s already taken a non-deterministic branch anyway. After all, we just care that the majority of validators reach the same result and preserve consensus. So worst-case scenario, there’s a dispute, and we can always admit fault and refund the wrong validator. On the other hand, if all validators take the code path that results in a seccomp violation, then they would all vote against the current candidate, which is also fine. The violation would get logged (in big scary letters) and hopefully some validator reports it to us.

Actually, a worst-worse-case scenario is that 50% of validators vote against, so that there is no consensus. But so many things would have to go wrong for that to happen:

An update to wasmtime is introduced that uses io_uring (unlikely as io_uring is mainly for IO-heavy applications)

The new syscall is not detected by our static analysis

It is never triggered in any of our tests

It then gets triggered on some super edge case in production on 50% of validators causing a stall (bad but very unlikely)

Or, it triggers on only a few validators causing a dispute (more likely but not as bad)

Considering how many things would have to go wrong here, we believe it’s safe to block io_uring.

Action on syscall violations

On syscall violations we currently only log, to make sure this works correctly before enforcing.

In the future, when a forbidden syscall is attempted we immediately kill the process in order to prevent the attacker from doing anything else. In execution, this will result in voting against the candidate.

Enums

Error

Functions

check_is_fully_enabled

Runs a check for seccomp and returns a single bool indicating whether seccomp with our rules is fully enabled on the current Linux environment.

enable_for_worker

Try to enable seccomp for the given kind of worker.

try_restrict

Applies a seccomp filter to disable networking for the PVF threads.

Type Definitions

Result