Crate polkadot_node_core_pvf_common

Contains functionality related to PVFs that is shared by the PVF host and the PVF workers.

Re-exports

pub use sp_tracing;

Modules

error

execute

executor_interface

Interface to the Substrate Executor

prepare

pvf

tests

worker

Functionality common to both prepare and execute workers.

worker_dir

Shared functions for getting the known worker files.

Macros

decl_worker_main

Use this macro to declare a fn main() {} that will create an executable that can be used for spawning the desired worker.

Structs

ProcessTime

CPU Time Used by The Whole Process

SecurityStatus

Status of security features on the current system.

WorkerHandshake

A handshake with information for the worker.

Functions

framed_recv_blocking

Read some data prefixed by its length from r. Sync version of framed_recv to avoid dependency on tokio.

framed_send_blocking

Write some data prefixed by its length into w. Sync version of framed_send to avoid dependency on tokio.