Module polkadot_node_core_pvf_common::worker::security::landlock

The landlock docs say it best:

“Landlock is a security feature available since Linux 5.13. The goal is to enable to restrict ambient rights (e.g., global filesystem access) for a set of processes by creating safe security sandboxes as new security layers in addition to the existing system-wide access-controls. This kind of sandbox is expected to help mitigate the security impact of bugs, unexpected or malicious behaviors in applications. Landlock empowers any process, including unprivileged ones, to securely restrict themselves.”

Enums

Error

RulesetStatus

Enforcement status of a ruleset.

Constants

LANDLOCK_ABI

Landlock ABI version. We use ABI V1 because:

Functions

check_can_fully_enable

Runs a check for landlock in its own thread, and returns an error indicating whether the given landlock ABI is fully enabled on the current Linux environment.

enable_for_worker

Try to enable landlock for the given kind of worker.

Type Aliases

Result